# BFGUI API Documentation/BFGUI_Api_Documentation.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BFGUI Api Documentation",
    "version": "v0.39.3"
  },
  "servers": [
    {
      "url": "https://bfgui-dev.webdev.banksvcs.net/bfgui",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/transactions": {
      "post": {
        "tags": [
          "sct-transaction-search-controller"
        ],
        "operationId": "getSCTTransactions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionSearchCriteria"
              }
            }
          }
        },
        "description": "Returns the page with SCT Transactions filtered by search fields",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJURVNUVVNFUjIiLCJwZXJtaXNzaW9ucyI6WyJTRkdfU0NUX0lDRk91dGJvdW5kIiwiU0ZHX1VJX1NDVF9FTlRJVFlfVFJBTlNNSVQiLCJTRkdfVUlfU0NUX0RFTEVURV9FTlRJVFlfU0NUIiwiRkJfVUlfVFJVU1RFRF9DRVJUU19ORVciLCJTRkdfVUlfSE9NRSIsIlNGR19VSV9TQ1RfQVBQUk9WRV9FTlRJVFlfR1BMIiwiU0ZHX1VJX1NDVF9FRElUX0VOVElUWV9TQ1QiLCJTRkdfVUlfU0NUX0RFTEVURV9FTlRJVFlfR1BMIiwiRkJfVUlfVFJVU1RFRF9DRVJUU19ERUxFVEUiLCJTRkdfVUlfU0NUX0NSRUFURV9FTlRJVFlfU0NUIiwiU0ZHX1VJX1NDVF9FTlRJVFkiLCJGQl9VSV9UUlVTVEVEX0NFUlRTX0FQUFJPVkUiLCJTRkdfVUlfU0NUX0FQUFJPVkVfRU5USVRZX1NDVCIsIlNGR19VSV9TQ1RfRURJVF9FTlRJVFlfR1BMIiwiU0ZHX1VJX1NDVF9DUkVBVEVfRU5USVRZX0dQTCIsIkZCX1VJX1RSVVNURURfQ0VSVFMiXSwiYWF0IjoxMCwidGV0IjoxMjAsImlhdCI6MTYzMzk1NjI5NiwiZXhwIjoxNjMzOTYzNDk2fQ.4vJ6iy7P5WD_mIsRHb3a66UOJOGVzZ-FzMnFxCwf8qxO3kHLJwvK3JC9U3k9ab-HNTMI7-31qx6t4ckGSFQBVQ"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTransaction"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/auth/sso": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "signInSso",
        "description": "Returns JWT token, authenticates the user.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginSsoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/auth/signin": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "signIn",
        "description": "Returns JWT, authenticates the user.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/auth/reauth": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "continueSession",
        "description": "Returns JWT token to continue session.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/transactions/{id}": {
      "get": {
        "tags": [
          "sct-transaction-search-controller"
        ],
        "operationId": "getTransactionById",
        "description": "Returns SCT Transaction data by ID.",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJURVNUVVNFUjIiLCJwZXJtaXNzaW9ucyI6WyJTRkdfU0NUX0lDRk91dGJvdW5kIiwiU0ZHX1VJX1NDVF9FTlRJVFlfVFJBTlNNSVQiLCJTRkdfVUlfU0NUX0RFTEVURV9FTlRJVFlfU0NUIiwiRkJfVUlfVFJVU1RFRF9DRVJUU19ORVciLCJTRkdfVUlfSE9NRSIsIlNGR19VSV9TQ1RfQVBQUk9WRV9FTlRJVFlfR1BMIiwiU0ZHX1VJX1NDVF9FRElUX0VOVElUWV9TQ1QiLCJTRkdfVUlfU0NUX0RFTEVURV9FTlRJVFlfR1BMIiwiRkJfVUlfVFJVU1RFRF9DRVJUU19ERUxFVEUiLCJTRkdfVUlfU0NUX0NSRUFURV9FTlRJVFlfU0NUIiwiU0ZHX1VJX1NDVF9FTlRJVFkiLCJGQl9VSV9UUlVTVEVEX0NFUlRTX0FQUFJPVkUiLCJTRkdfVUlfU0NUX0FQUFJPVkVfRU5USVRZX1NDVCIsIlNGR19VSV9TQ1RfRURJVF9FTlRJVFlfR1BMIiwiU0ZHX1VJX1NDVF9DUkVBVEVfRU5USVRZX0dQTCIsIkZCX1VJX1RSVVNURURfQ0VSVFMiXSwiYWF0IjoxMCwidGV0IjoxMjAsImlhdCI6MTYzMzk1NjI5NiwiZXhwIjoxNjMzOTYzNDk2fQ.4vJ6iy7P5WD_mIsRHb3a66UOJOGVzZ-FzMnFxCwf8qxO3kHLJwvK3JC9U3k9ab-HNTMI7-31qx6t4ckGSFQBVQ"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/transactions/transaction-criteria-data": {
      "get": {
        "tags": [
          "sct-transaction-search-controller"
        ],
        "operationId": "getTransactionCriteriaData",
        "description": "Returns SCT transaction criteria data for searching sct transactions.",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJURVNUVVNFUjIiLCJwZXJtaXNzaW9ucyI6WyJTRkdfU0NUX0lDRk91dGJvdW5kIiwiU0ZHX1VJX1NDVF9FTlRJVFlfVFJBTlNNSVQiLCJTRkdfVUlfU0NUX0RFTEVURV9FTlRJVFlfU0NUIiwiRkJfVUlfVFJVU1RFRF9DRVJUU19ORVciLCJTRkdfVUlfSE9NRSIsIlNGR19VSV9TQ1RfQVBQUk9WRV9FTlRJVFlfR1BMIiwiU0ZHX1VJX1NDVF9FRElUX0VOVElUWV9TQ1QiLCJTRkdfVUlfU0NUX0RFTEVURV9FTlRJVFlfR1BMIiwiRkJfVUlfVFJVU1RFRF9DRVJUU19ERUxFVEUiLCJTRkdfVUlfU0NUX0NSRUFURV9FTlRJVFlfU0NUIiwiU0ZHX1VJX1NDVF9FTlRJVFkiLCJGQl9VSV9UUlVTVEVEX0NFUlRTX0FQUFJPVkUiLCJTRkdfVUlfU0NUX0FQUFJPVkVfRU5USVRZX1NDVCIsIlNGR19VSV9TQ1RfRURJVF9FTlRJVFlfR1BMIiwiU0ZHX1VJX1NDVF9DUkVBVEVfRU5USVRZX0dQTCIsIkZCX1VJX1RSVVNURURfQ0VSVFMiXSwiYWF0IjoxMCwidGV0IjoxMjAsImlhdCI6MTYzMzk1NjI5NiwiZXhwIjoxNjMzOTYzNDk2fQ.4vJ6iy7P5WD_mIsRHb3a66UOJOGVzZ-FzMnFxCwf8qxO3kHLJwvK3JC9U3k9ab-HNTMI7-31qx6t4ckGSFQBVQ"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TransactionSearchCriteria": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "wfid": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "writeOnly": true
          },
          "start": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "entity": {
            "type": "string"
          },
          "paymentbic": {
            "type": "string"
          },
          "transactionid": {
            "type": "string"
          },
          "payaway": {
            "type": "boolean"
          },
          "rows": {
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "is-error": {
            "type": "boolean"
          },
          "exclude-archive": {
            "type": "boolean"
          },
          "settlement-from": {
            "type": "string"
          },
          "settlement-to": {
            "type": "string"
          },
          "ptimestamp-from": {
            "type": "string"
          },
          "ptimestamp-to": {
            "type": "string"
          },
          "has-message-id": {
            "type": "boolean"
          },
          "transaction-type": {
            "type": "string",
            "enum": [
              "transaction",
              "payaway"
            ]
          }
        }
      },
      "PageTransaction": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/Sort"
          },
          "pageable": {
            "$ref": "#/components/schemas/Pageable"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/Sort"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Sort": {
        "type": "object",
        "properties": {
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "transactionID": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "settleDate": {
            "type": "string",
            "format": "date-time"
          },
          "settleAmount": {
            "type": "number",
            "format": "double"
          },
          "workflowID": {
            "type": "integer",
            "format": "int32"
          },
          "direction": {
            "type": "string"
          },
          "fileID": {
            "type": "string"
          },
          "docID": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "payaway": {
            "type": "boolean"
          }
        }
      },
...
      "LoginSsoRequest": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "nodeName": {
            "type": "string"
          },
          "dlssoToken": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    }
  }
}